Improve tools registration (#37)
* Adding a new method for registering tools that allows MCP servers to
actively post to the ChARGe server that they are available.

* Refactored how the charge backend MCP servers are invoked to allow
them to register themselves with the backend server.

* Removed deprecated CLI arguments.

* Changed the servers global list to be maintained as a dictionary.
This enables a unique server, port mapping that can be overwritten or
evicted.

* Added support to check each MCP server URL to make sure that it is
active.  Stale servers are removed.

Cleaned up the callback logger to fix the class style, allow it to
unbind when the websocket is closed and removed the async part of the
logging to the UI.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

* Address reviewer feedback.

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>