Improve multi-metric LMO (#95)
* Code cleanup.  Added new code to only add a metric if it has changed.

* Made the property prediction and comparison more robust for all property types.  Improved logging to copilot.

* Added support to extract the property being optimized in LMO into the
reasoning result as part of the structured output.  This value is then
captured in the LMO backend and is saved to the node state.  Note that
any tool response is favored over a direct function available to the
LMO backend.  At this time, only a single property per query is
properly saved in the structured output.

* Fixed how the initial molecule properties are selected.  Added a note that the LMO task needs to be made to take a function for the property.  Fixed formatting in the hover-over window.

* Apply suggestions from code review

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>