Connect available tools (#117)
* Added the list of currently available tools to the task manager so
that it is forwarded to the generated tasks.

* Improved how error handling is done for tasks and added messaging back
to the UI reasoning window.  Improved how the _handle_task_done
function deals with exceptions and avoids raising an exception in a
background async task, since there is nothing left to catch it. This
eliminated some runtime warning messages.  Fixed the name for the
logging of generate_molecule to be singular.

* Update the retro synthesis tasks to use the avaiable tools selected by
the user.

* Added a feature to the callback logger to allow a message source to be
set when initialized.

* Improving the logging within the retrosynthesis task.

* Passed the available tools to the retrosynthesis tasks.  Added
improved logging for retrosynthesis.